+++ services/security/vault-raft/CHANGELOG.md
@@ -122,3 +122,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - initial docker files (9d783148)
+
+- add policies mount (e76bca9c)